Update package.json
The initial package.json file was not installing properly. Used the file from https://github.com/shine7653/Homework12-EmployeeTracker to get a functional install. The functionality of the code is a result of office hours with Tim & Josh (TA's) as well as Ben (tutor) to create a working program with all bonus functionality. After speaking with Lisa (SSM) Made modifications to the file and sited the source in the commit changes to avoid plagiarism and give credit where credit is due.